Holding update for kdeconnect-frameworks-git #77
kdeconnect-frameworks-git was depended on kcmutils, until recently an update
of upstream added the dependency for APIs provided in kcmutils-git. The latter
package will in turn pull in much more kde -git dependencies including
kdeclarative-git kio-git kwallet-git karchive-git kbookmarks-git
kjobwidgets-git solid-git (, and possibly more of them).

Pulling and replacing so many -git packages in this repo is not an ideal
solution, therefore I decide to hold the update for kdeconnect-frameworks-git
for the foreseeable future, until:
1. Official extra/kcmutils update to support APIs used in
   kdeconnect-frameworks-git, or
2. kdeconnect-frameworks officially release and replace the -git version.

Meanwhile I will look for other workarounds to make kdeconnect work and keep
updating again under KDE5, any suggestions are welcome.
This issue will be closed as soon as a solution/workaround appears.