Revert to the commit before the destructive 7bd463557ffd769b7bdafe13676b418696c435cc by @petronny
This should restore the broken ARM and AArch64 support.

This also explains why the versioning was broken a few weeks ago
and possibly why the packages was removed from the repo too.

The server URL didn't change, the new URL redirects to the old one.
Switch to git protocol in case that's more reliable on their server.