aurman: new package
As per [1], trizen, aurman and yay are three most popular AUR helpers.
We have trizen and yay, so why not aurman :)

For aurman, new releases are published for almost every single git
commit according to recent git history. As a result, aurman-git is
in general as new as aurman, so I didn't add aurman-git.

[1] https://aur.archlinux.org/packages/ (fetched 2018.05.29)