libpdfium-nojs: force to call prepare() in pkgver()
This is required because archlinuxcn/lilac#124 added the --noprepare
flag when calling makepkg, but this package finds and checks out the
latest version in prepare() before using pkgver() to print that version.

This is a bit hacky but it works... Probably it's better to revert the
aforementioned pull request as PKGBUILD's documentation explicitly
states that pkgver() is called after prepare().