screengrab-git: update the AUR package from this repo
To address #2228, ykelvis added the `maintainers` argument to
`aur_pre_build` on 5/11 14:41 CST [1]. However, coincidentally,
jleclanche orphaned most of his LXQt packages on 5/12 07:24 CST, so the
last update fails. As I (yan12125) am now the only maintainer of the AUR
package, I switch the update model to make this repo the source of
trust.

[1] commit 8a7b037d63e58fce86377b3a83abf9c881071009